signaling/requirements.txt
fastapi
uvicorn 
python-socketio
aiohttp